+++ database/migrations/20200714184243_add_notify_and_extensions_to_conferences_table.php
@@ -8,9 +8,7 @@ class AddNotifyAndExtensionsToConferencesTable extends Migration
 {
     public function up()
     {
-        $this->disableForeignKeyCheck();
-        Info::truncate();
-        $this->enableForeignKeyCheck();
+        Info::query()->delete();
 
         $this->schema->table('conferences', function (Blueprint $table) {
             $table->text('extensions')->nullable();